+++ test/examples/gms/bearing.gms
