+++ test/examples/gms/camcge.gms
