+++ test/examples/gms/ex14_2_4.gms
@@ -1,0 +1,115 @@
+$offlisting
+*  
+*  Equation counts
+*      Total        E        G        L        N        X        C        B
+*          8        2        0        6        0        0        0        0
+*  
+*  Variable counts
+*                   x        b        i      s1s      s2s       sc       si
+*      Total     cont   binary  integer     sos1     sos2    scont     sint
+*          6        6        0        0        0        0        0        0
+*  FX      0
+*  
+*  Nonzero counts
+*      Total    const       NL      DLL
+*         35       11       24        0
+*
+*  Solve m using NLP minimizing objvar;
+
+
+Variables  x1,x2,x3,x4,objvar,x6;
+
+Positive Variables  x6;
+
+Equations  e1,e2,e3,e4,e5,e6,e7,e8;
+
+
+e1..    objvar - x6 =E= 0;
+
+e2.. (0.549337520233386*x2 + 1.1263896788319*x3)/(x1 + 0.816722116903399*x2 + 
+     0.538540530229217*x3) + 0.0910522583583458*x2/(0.972203312166101*x1 + x2
+      + 0.394821041898112*x3) - 0.273994101407968*x3/(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3) - (x1*(0.549337520233386*x2 + 1.1263896788319*
+     x3)/sqr(x1 + 0.816722116903399*x2 + 0.538540530229217*x3) + 
+     0.972203312166101*x2*(0.0910522583583458*x1 + 1.03765878646318*x3)/sqr(
+     0.972203312166101*x1 + x2 + 0.394821041898112*x3) + 1.07810138009609*x3*(
+     0.692718766203089*x2 - 0.273994101407968*x1)/sqr(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3)) - 3667.70490156687/(226.184 + x4) - x6
+      =L= -12.0457123581059;
+
+e3.. (0.0910522583583458*x1 + 1.03765878646318*x3)/(0.972203312166101*x1 + x2
+      + 0.394821041898112*x3) + 0.549337520233386*x1/(x1 + 0.816722116903399*x2
+      + 0.538540530229217*x3) + 0.692718766203089*x3/(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3) - (0.816722116903399*x1*(0.549337520233386*x2
+      + 1.1263896788319*x3)/sqr(x1 + 0.816722116903399*x2 + 0.538540530229217*
+     x3) + x2*(0.0910522583583458*x1 + 1.03765878646318*x3)/sqr(
+     0.972203312166101*x1 + x2 + 0.394821041898112*x3) + 0.707289137797622*x3*(
+     0.692718766203089*x2 - 0.273994101407968*x1)/sqr(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3)) - 2904.34268119711/(221.969 + x4) - x6
+      =L= -9.63112952618865;
+
+e4.. (0.692718766203089*x2 - 0.273994101407968*x1)/(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3) + 1.1263896788319*x1/(x1 + 0.816722116903399*x2
+      + 0.538540530229217*x3) + 1.03765878646318*x2/(0.972203312166101*x1 + x2
+      + 0.394821041898112*x3) - (0.538540530229217*x1*(0.549337520233386*x2 + 
+     1.1263896788319*x3)/sqr(x1 + 0.816722116903399*x2 + 0.538540530229217*x3)
+      + 0.394821041898112*x2*(0.0910522583583458*x1 + 1.03765878646318*x3)/sqr(
+     0.972203312166101*x1 + x2 + 0.394821041898112*x3) + x3*(0.692718766203089*
+     x2 - 0.273994101407968*x1)/sqr(1.07810138009609*x1 + 0.707289137797622*x2
+      + x3)) - 3984.92283948829/(233.426 + x4) - x6 =L= -11.9515596536534;
+
+e5.. (-(0.549337520233386*x2 + 1.1263896788319*x3)/(x1 + 0.816722116903399*x2
+      + 0.538540530229217*x3)) - (0.0910522583583458*x2/(0.972203312166101*x1
+      + x2 + 0.394821041898112*x3) - 0.273994101407968*x3/(1.07810138009609*x1
+      + 0.707289137797622*x2 + x3)) + x1*(0.549337520233386*x2 + 
+     1.1263896788319*x3)/sqr(x1 + 0.816722116903399*x2 + 0.538540530229217*x3)
+      + 0.972203312166101*x2*(0.0910522583583458*x1 + 1.03765878646318*x3)/sqr(
+     0.972203312166101*x1 + x2 + 0.394821041898112*x3) + 1.07810138009609*x3*(
+     0.692718766203089*x2 - 0.273994101407968*x1)/sqr(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3) + 3667.70490156687/(226.184 + x4) - x6
+      =L= 12.0457123581059;
+
+e6.. (-(0.0910522583583458*x1 + 1.03765878646318*x3)/(0.972203312166101*x1 + x2
+      + 0.394821041898112*x3)) - (0.549337520233386*x1/(x1 + 0.816722116903399*
+     x2 + 0.538540530229217*x3) + 0.692718766203089*x3/(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3)) + 0.816722116903399*x1*(0.549337520233386*x2
+      + 1.1263896788319*x3)/sqr(x1 + 0.816722116903399*x2 + 0.538540530229217*
+     x3) + x2*(0.0910522583583458*x1 + 1.03765878646318*x3)/sqr(
+     0.972203312166101*x1 + x2 + 0.394821041898112*x3) + 0.707289137797622*x3*(
+     0.692718766203089*x2 - 0.273994101407968*x1)/sqr(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3) + 2904.34268119711/(221.969 + x4) - x6
+      =L= 9.63112952618865;
+
+e7.. (-(0.692718766203089*x2 - 0.273994101407968*x1)/(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3)) - (1.1263896788319*x1/(x1 + 0.816722116903399*
+     x2 + 0.538540530229217*x3) + 1.03765878646318*x2/(0.972203312166101*x1 + 
+     x2 + 0.394821041898112*x3)) + 0.538540530229217*x1*(0.549337520233386*x2
+      + 1.1263896788319*x3)/sqr(x1 + 0.816722116903399*x2 + 0.538540530229217*
+     x3) + 0.394821041898112*x2*(0.0910522583583458*x1 + 1.03765878646318*x3)/
+     sqr(0.972203312166101*x1 + x2 + 0.394821041898112*x3) + x3*(
+     0.692718766203089*x2 - 0.273994101407968*x1)/sqr(1.07810138009609*x1 + 
+     0.707289137797622*x2 + x3) + 3984.92283948829/(233.426 + x4) - x6
+      =L= 11.9515596536534;
+
+e8..    x1 + x2 + x3 =E= 1;
+
+* set non-default bounds
+x1.lo = 1E-6; x1.up = 1;
+x2.lo = 1E-6; x2.up = 1;
+x3.lo = 1E-6; x3.up = 1;
+x4.lo = 40; x4.up = 90;
+
+* set non-default levels
+x1.l = 0.187;
+x2.l = 0.56;
+x3.l = 0.253;
+x4.l = 72.957;
+
+Model m / all /;
+
+m.limrow=0; m.limcol=0;
+
+$if NOT '%gams.u1%' == '' $include '%gams.u1%'
+
+$if not set NLP $set NLP NLP
+Solve m using %NLP% minimizing objvar;
