+++ test/examples/gms/ex14_2_8.gms
@@ -1,0 +1,82 @@
+*  NLP written by GAMS Convert at 07/19/01 13:40:30
+*  
+*  Equation counts
+*     Total       E       G       L       N       X
+*         6       2       0       4       0       0
+*  
+*  Variable counts
+*                 x       b       i     s1s     s2s      sc      si
+*     Total    cont  binary integer    sos1    sos2   scont    sint
+*         5       5       0       0       0       0       0       0
+*  FX     0       0       0       0       0       0       0       0
+*  
+*  Nonzero counts
+*     Total   const      NL     DLL
+*        20       8      12       0
+*
+*  Solve m using NLP minimizing objvar;
+
+
+Variables  x1,x2,x3,objvar,x5;
+
+Positive Variables x5;
+
+Equations  e1,e2,e3,e4,e5,e6;
+
+
+e1..    objvar - x5 =E= 0;
+
+e2.. 10.68*log(2.5735*x1 + 4.0464*x2) - 9.344*log(2.336*x1 + 3.24*x2) - (
+     2.5364416*x2 - 0.993370999999997*x1)/(2.5735*x1 + 4.0464*x2) - (1.696*log(
+     1.69610217540928*x1 + 3.24*x2) + 0.64*log(0.657731453039811*x1 + 
+     0.0338737664203932*x2)) - (2.87658928949414*x1/(1.69610217540928*x1 + 3.24
+     *x2) + 5.49537104832607*x2/(1.69610217540928*x1 + 3.24*x2) + 
+     0.420948129945479*x1/(0.657731453039811*x1 + 0.0338737664203932*x2)) - 
+     2787.49800065313/(229.664 + x3) - x5 =L= -10.164795069335;
+
+e3.. 15.2*log(2.5735*x1 + 4.0464*x2) - 12.96*log(2.336*x1 + 3.24*x2) - (
+     3.98813184*x2 - 1.5619104*x1)/(2.5735*x1 + 4.0464*x2) - 3.24*log(
+     1.69610217540928*x1 + 3.24*x2) - (5.49504*x1/(1.69610217540928*x1 + 3.24*
+     x2) + 10.4976*x2/(1.69610217540928*x1 + 3.24*x2) + 0.0216792105090516*x1/(
+     0.657731453039811*x1 + 0.0338737664203932*x2)) - 2766.63/(222.65 + x3)
+      - x5 =L= -11.1422900361581;
+
+e4.. 9.344*log(2.336*x1 + 3.24*x2) - 10.68*log(2.5735*x1 + 4.0464*x2) + (
+     2.5364416*x2 - 0.993370999999997*x1)/(2.5735*x1 + 4.0464*x2) + 1.696*log(
+     1.69610217540928*x1 + 3.24*x2) + 0.64*log(0.657731453039811*x1 + 
+     0.0338737664203932*x2) + 2.87658928949414*x1/(1.69610217540928*x1 + 3.24*
+     x2) + 5.49537104832607*x2/(1.69610217540928*x1 + 3.24*x2) + 
+     0.420948129945479*x1/(0.657731453039811*x1 + 0.0338737664203932*x2) + 
+     2787.49800065313/(229.664 + x3) - x5 =L= 10.164795069335;
+
+e5.. 12.96*log(2.336*x1 + 3.24*x2) - 15.2*log(2.5735*x1 + 4.0464*x2) + (
+     3.98813184*x2 - 1.5619104*x1)/(2.5735*x1 + 4.0464*x2) + 3.24*log(
+     1.69610217540928*x1 + 3.24*x2) + 5.49504*x1/(1.69610217540928*x1 + 3.24*x2
+     ) + 10.4976*x2/(1.69610217540928*x1 + 3.24*x2) + 0.0216792105090516*x1/(
+     0.657731453039811*x1 + 0.0338737664203932*x2) + 2766.63/(222.65 + x3) - x5
+      =L= 11.1422900361581;
+
+e6..    x1 + x2 =E= 1;
+
+* set non default bounds
+
+x1.lo = 1E-6; x1.up = 1; 
+x2.lo = 1E-6; x2.up = 1; 
+x3.lo = 40; x3.up = 90; 
+
+* set non default levels
+
+x1.l = 0.878; 
+x2.l = 0.122; 
+x3.l = 55.726; 
+
+* set non default marginals
+
+
+Model m / all /;
+
+m.limrow=0; m.limcol=0;
+
+$if NOT '%gams.u1%' == '' $include '%gams.u1%'
+
+Solve m using NLP minimizing objvar;
